backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
python-dotenv>=1.0.1

# NLP / Vector Search (MVP)
numpy>=1.26.0
faiss-cpu>=1.8.0

# Default embedder fallback (lightweight)
scikit-learn>=1.4.0

